refactor: reorganise code; move reconsile logic to reconsile.go
